chore: Remove group unification in rows
Rather than all groups of a unified group be collapsed into a
comma-separated list, the id of a new group will be placed as a separate
column. This was changed because reading in the collapsed csv groups
took too long. You'll have to sort the new groups manually in downstream
analysis